+++ src/lib.rs
@@ -1,7 +1,2 @@
-pub mod token;
-pub mod lexer;
-pub mod ast;
-pub mod parser;
+pub mod frontend;
 pub mod eval;
-
-mod macros;
